requirements.txt
requests
aiohttp
python-telegram-bot
flask
scikit-learn
pandas
joblib
